cast to float

+++ numpy_version/vectorization.py
@@ -15,16 +15,16 @@ def speed():
     c = np.zeros(n)
     for i in range(n):
         c[i] = a[i] + b[i]
-    t1 = timer.stop()
+    t1 = float(timer.stop())
 
     timer.start()
     d = a + b
-    t2 = timer.stop()
+    t2 = float(timer.stop())
     print("d size {}".format(d.size))
     print('for loop : %.5f sec' % t1)
     print('vectors : %.10f sec' % t2)
     if t2 > 0:
-        print('ratio : %.1f sec' % t1 / t2)
+        print('ratio : %.1f sec' % float(t1 / t2))
 
 
 def normal(z, mu, sigma):
